Add dashboard page for viewing the job queue
In general it works, but the solution using the two consecutive SF API
calls is not ideal (first to run sacct, second to get task output).
The second one to get the task output runs too quickly unless a sleep
is inserted.

Using jq to filter the JSON output of sacct. Added to requirements.txt