Change how MongoDB settings are read in start_runner
The MongoDB settings now attempt to read from environment variables with
default values if unset.

The script first now tries to source a mongo_env.sh script to load the
settings. This way the Mongo user/pass URL can be stored in a file and
read in and kept separate from files uploaded to Github.

Added the sbatch_runner.sh script for submitting the job.