Add new pages to show profiling data stored in the DB
Similar to the webhooks pages, the new profile pages show an overview of
the stored profiles in the DB and have a separate page for viewing the
details of a single profile.

New script to upload Nsys CSV reports into the DB.

Having everything in a single app.py file is getting unwieldy due to its
length. Should reorganize into a few files.